Extract common configuration code to parent class
